have working models with correct migrations. formatted models
