+++ website/forms/__init__.py
@@ -1,0 +1,1 @@
+from .forms import UserForm, ProductForm
