+++ website/views/__init__.py
@@ -1,0 +1,1 @@
+from .views import index, register, login_user, user_logout, sell_product, list_products
